fix(trusty-memory): address #2522 review — split startup_task_tests (drop allowlist entry) + log multi-tenant activation
Refs #2319 #2223
